+++ src/com/atteo/jello/store/Page.java
@@ -8,7 +8,7 @@ import com.google.inject.name.Named;
 public class Page {
 	public Page nextInPool = null; // public for performance reasons
 	protected int accessCount = 0;
-	protected long id;
+	protected int id;
 	protected byte data[] = null;
 	protected byte accessibleData[] = null;
 	protected boolean dirty = false;
@@ -16,7 +16,7 @@ public class Page {
 	
 	@Inject
 	public
-	Page(@Named("pageSize") final int pageSize) {
+	Page(@Named("pageSize") final short pageSize) {
 		data = new byte[pageSize];
 		byteBuffer = ByteBuffer.wrap(data);
 		byteBuffer.position(headerSize());
@@ -47,19 +47,19 @@ public class Page {
 		this.dirty = dirty;
 	}
 
-	public int getCapacity() {
-		return accessibleData.length;
+	public short getCapacity() {
+		return (short) accessibleData.length;
 	}
 	
-	public int headerSize() {
+	public short headerSize() {
 		return 0;
 	}
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 	
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 }
